feat: add user public list (#6)
* feat: add user public list

* fix: firestore rules

* feat: allow change data before adding to user list

+++ src/containers/Anime.tsx
@@ -3,15 +3,17 @@ import { AnimeModel, Score, Status } from '../models';
 import firebase from '../firebase/firebase';
 
 type ComponentProps = {
+  editable: boolean;
   onChange: (value: string | undefined | null | Status | Score, propname: string) => void;
   onDelete: () => void;
 } & AnimeModel;
 
 type Props = {
+  editable: boolean;
   component: FunctionComponent<ComponentProps>;
 } & AnimeModel;
 
-export const Anime: React.FC<Props> = ({ component, id, ...rest }) => {
+export const Anime: React.FC<Props> = ({ component, id, editable, ...rest }) => {
   const docRef = useMemo(() => 
     firebase
       .firestore()
@@ -23,11 +25,13 @@ export const Anime: React.FC<Props> = ({ component, id, ...rest }) => {
     value: string | undefined | null | Status | Score | string[],
     propname: string
   ) => {
-    docRef.update({ [propname]: value });
+    if (editable)
+      docRef.update({ [propname]: value });
   };
   const onDelete = () => {
-    docRef.delete();
+    if (editable)
+      docRef.delete();
   };
 
-  return React.createElement(component, { id, onChange, onDelete, ...rest });
+  return React.createElement(component, { id, editable, onChange, onDelete, ...rest });
 };
